Saca Duty Cycle

+++ BCU-Shared-H10/Inc/Shared/SPI.hpp
@@ -10,28 +10,27 @@ SPIStackOrder *create_state_order(StateMachine::state_id *master_general_state,
                                   StateMachine::state_id *slave_nested_state);
 
 SPIStackOrder *create_control_parameters_order(
-    double *velocity_reference, double *velocity_error,
-    double *u_current_measurement, double *v_current_measurement,
-    double *w_current_measurement, double *electrical_angle,
-    double *d_current_reference, double *d_current_measurement,
-    double *d_current_error, double *q_current_reference,
-    double *q_current_measurement, double *q_current_error,
-    double *three_phase_unbalance, double *d_target_voltage,
-    double *q_target_voltage, double *u_target_voltage,
-    double *v_target_voltage, double *w_target_voltage,
-    double *u_output_voltage, double *v_output_voltage,
-    double *w_output_voltage, double *u_duty_cycle, double *v_duty_cycle,
-    double *w_duty_cycle, double *angular_velocity);
+    float *velocity_reference, float *velocity_error,
+    float *u_current_measurement, float *v_current_measurement,
+    float *w_current_measurement, float *electrical_angle,
+    float *d_current_reference, float *d_current_measurement,
+    float *d_current_error, float *q_current_reference,
+    float *q_current_measurement, float *q_current_error,
+    float *three_phase_unbalance, float *d_target_voltage,
+    float *q_target_voltage, float *u_target_voltage, float *v_target_voltage,
+    float *w_target_voltage, float *u_output_voltage, float *v_output_voltage,
+    float *w_output_voltage, float *u_duty_cycle, float *v_duty_cycle,
+    float *w_duty_cycle, float *angular_velocity);
 
-SPIStackOrder *create_start_velocity_control_order(double *velocity_reference);
-SPIStackOrder *create_start_current_control_order(double *d_current_reference,
-                                                  double *q_current_reference);
-SPIStackOrder *create_start_emulated_movement_order(double *d_current_reference,
-                                                    double *q_current_reference,
-                                                    double *angular_velocity);
-SPIStackOrder *create_start_test_pwm_order(double *duty_cycle_u,
-                                           double *duty_cycle_v,
-                                           double *duty_cycle_w);
+SPIStackOrder *create_start_velocity_control_order(float *velocity_reference);
+SPIStackOrder *create_start_current_control_order(float *d_current_reference,
+                                                  float *q_current_reference);
+SPIStackOrder *create_start_emulated_movement_order(float *d_current_reference,
+                                                    float *q_current_reference,
+                                                    float *angular_velocity);
+SPIStackOrder *create_start_test_pwm_order(float *duty_cycle_u,
+                                           float *duty_cycle_v,
+                                           float *duty_cycle_w);
 
 SPIStackOrder *create_stop_control_order();
 
@@ -43,9 +42,9 @@ enum Direction : uint8_t {
 };
 
 SPIStackOrder *create_position_encoder_order(
-    std::array<double *, 3> position, std::array<double *, 3> velocity,
-    std::array<double *, 3> acceleration, std::array<Direction *, 3> direction,
-    double *average_position, double *max_velocity, bool *is_detecting);
+    std::array<float *, 3> position, std::array<float *, 3> velocity,
+    std::array<float *, 3> acceleration, std::array<Direction *, 3> direction,
+    float *average_position, float *max_velocity, bool *is_detecting);
 
 SPIStackOrder *create_force_dc_link_order(float *dc_link_voltage);
 SPIStackOrder *create_unlock_dc_link_order();
