+++ BCU-Shared-H10/Src/Shared/SPI.cpp
@@ -17,24 +17,23 @@ SPIStackOrder *create_state_order(StateMachine::state_id *master_general_state,
 }
 
 SPIStackOrder *create_control_parameters_order(
-    double *velocity_reference, double *velocity_error,
-    double *u_current_measurement, double *v_current_measurement,
-    double *w_current_measurement, double *electrical_angle,
-    double *d_current_reference, double *d_current_measurement,
-    double *d_current_error, double *q_current_reference,
-    double *q_current_measurement, double *q_current_error,
-    double *three_phase_unbalance, double *d_target_voltage,
-    double *q_target_voltage, double *u_target_voltage,
-    double *v_target_voltage, double *w_target_voltage,
-    double *u_output_voltage, double *v_output_voltage,
-    double *w_output_voltage, double *u_duty_cycle, double *v_duty_cycle,
-    double *w_duty_cycle, double *angular_velocity) {
+    float *velocity_reference, float *velocity_error,
+    float *u_current_measurement, float *v_current_measurement,
+    float *w_current_measurement, float *electrical_angle,
+    float *d_current_reference, float *d_current_measurement,
+    float *d_current_error, float *q_current_reference,
+    float *q_current_measurement, float *q_current_error,
+    float *three_phase_unbalance, float *d_target_voltage,
+    float *q_target_voltage, float *u_target_voltage, float *v_target_voltage,
+    float *w_target_voltage, float *u_output_voltage, float *v_output_voltage,
+    float *w_output_voltage, float *u_duty_cycle, float *v_duty_cycle,
+    float *w_duty_cycle, float *angular_velocity) {
     return new SPIStackOrder{
         998, *new SPIPacket<0>(),
-        *new SPIPacket<sizeof(double) * 25, double, double, double, double,
-                       double, double, double, double, double, double, double,
-                       double, double, double, double, double, double, double,
-                       double, double, double, double, double, double, double>(
+        *new SPIPacket<sizeof(float) * 25, float, float, float, float, float,
+                       float, float, float, float, float, float, float, float,
+                       float, float, float, float, float, float, float, float,
+                       float, float, float, float>(
             velocity_reference, velocity_error, u_current_measurement,
             v_current_measurement, w_current_measurement, electrical_angle,
             d_current_reference, d_current_measurement, d_current_error,
@@ -45,36 +44,36 @@ SPIStackOrder *create_control_parameters_order(
             v_duty_cycle, w_duty_cycle, angular_velocity)};
 };
 
-SPIStackOrder *create_start_velocity_control_order(double *velocity_reference) {
+SPIStackOrder *create_start_velocity_control_order(float *velocity_reference) {
     return new SPIStackOrder{
-        997, *new SPIPacket<sizeof(double), double>{velocity_reference},
+        997, *new SPIPacket<sizeof(float), float>{velocity_reference},
         *new SPIPacket<0>()};
 }
 
-SPIStackOrder *create_start_current_control_order(double *d_current_reference,
-                                                  double *q_current_reference) {
+SPIStackOrder *create_start_current_control_order(float *d_current_reference,
+                                                  float *q_current_reference) {
     return new SPIStackOrder{996,
-                             *new SPIPacket<sizeof(double) * 2, double, double>{
+                             *new SPIPacket<sizeof(float) * 2, float, float>{
                                  d_current_reference, q_current_reference},
                              *new SPIPacket<0>()};
 }
 
-SPIStackOrder *create_start_emulated_movement_order(double *d_current_reference,
-                                                    double *q_current_reference,
-                                                    double *angular_velocity) {
+SPIStackOrder *create_start_emulated_movement_order(float *d_current_reference,
+                                                    float *q_current_reference,
+                                                    float *angular_velocity) {
     return new SPIStackOrder{
         995,
-        *new SPIPacket<sizeof(double) * 3, double, double, double>{
+        *new SPIPacket<sizeof(float) * 3, float, float, float>{
             d_current_reference, q_current_reference, angular_velocity},
         *new SPIPacket<0>()};
 }
 
-SPIStackOrder *create_start_test_pwm_order(double *duty_cycle_u,
-                                           double *duty_cycle_v,
-                                           double *duty_cycle_w) {
+SPIStackOrder *create_start_test_pwm_order(float *duty_cycle_u,
+                                           float *duty_cycle_v,
+                                           float *duty_cycle_w) {
     return new SPIStackOrder{
         994,
-        *new SPIPacket<sizeof(double) * 3, double, double, double>{
+        *new SPIPacket<sizeof(float) * 3, float, float, float>{
             duty_cycle_u, duty_cycle_v, duty_cycle_w},
         *new SPIPacket<0>()};
 }
@@ -88,15 +87,15 @@ SPIStackOrder *create_enable_booster_order() {
 }
 
 SPIStackOrder *create_position_encoder_order(
-    std::array<double *, 3> position, std::array<double *, 3> velocity,
-    std::array<double *, 3> acceleration, std::array<Direction *, 3> direction,
-    double *average_position, double *max_velocity, bool *is_detecting) {
+    std::array<float *, 3> position, std::array<float *, 3> velocity,
+    std::array<float *, 3> acceleration, std::array<Direction *, 3> direction,
+    float *average_position, float *max_velocity, bool *is_detecting) {
     return new SPIStackOrder{
         991, *new SPIPacket<0>(),
         *new SPIPacket<
-            sizeof(double) * 12 + sizeof(Direction) * 3 + sizeof(bool), double,
-            double, double, double, double, double, double, double, double,
-            Direction, Direction, Direction, double, double, bool>(
+            sizeof(float) * 12 + sizeof(Direction) * 3 + sizeof(bool), float,
+            float, float, float, float, float, float, float, float, Direction,
+            Direction, Direction, float, float, bool>(
             position[0], position[1], position[2], velocity[0], velocity[1],
             velocity[2], acceleration[0], acceleration[1], acceleration[2],
             direction[0], direction[1], direction[2], average_position,
